+++ src/senaite/core/setuphandlers.py
@@ -255,6 +255,7 @@ def add_senaite_setup_items(portal):
         # ID, Title, FTI
         ("departments", "Lab Departments", "Departments"),
         ("labels", "Labels", "Labels"),
+        ("sampleconditions", "Sample Conditions", "SampleConditions"),
     ]
     setup = api.get_senaite_setup()
     add_dexterity_items(setup, items)